feat: confirm worker created successfully

+++ src/DFrame.KubernetesWorker/KubernetesScalingProvider.cs
@@ -1,19 +1,27 @@
 ﻿using System;
 using System.Linq;
 using System.Net.Http;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using DFrame.Core;
-using YamlDotNet.Serialization;
 
 namespace DFrame.KubernetesWorker
 {
     public class KubernetesScalingProvider : IScalingProvider
     {
-        private readonly string _ns = "dframe-worker";
-        private readonly string _name = "dframe";
         private readonly KubernetesApi _kubeapi;
-        private readonly IDeserializer _yamlDeserializer;
+        private readonly bool preserveWorker = true;
+
+        // master info
+        private readonly string masterSvc = "dframe-master";
+        private readonly string masterNamespace = "dframe";
+
+        // manifest configuable parameters
+        private readonly string _ns = "dframe";
+        private readonly string _name = "dframe-worker";
+        private readonly string imagePullSecret = "aws-registry";
+        private readonly string imagePullPolicy = "Never";
 
         private string _namespaceManifest;
         private string _deploymentManifest;
@@ -22,10 +30,9 @@ namespace DFrame.KubernetesWorker
         {
             _kubeapi = new KubernetesApi(new KubernetesApiConfig
             {
-                AccesptHeaderType = HeaderContentType.Yaml,
+                ResponseHeaderType = HeaderContentType.Yaml,
                 SkipCertificateValidation = true,
             });
-            _yamlDeserializer = new DeserializerBuilder().Build();
         }
 
         public async Task StartWorkerAsync(DFrameOptions options, int nodeCount, CancellationToken cancellationToken)
@@ -34,27 +41,28 @@ namespace DFrame.KubernetesWorker
             // todo: rbac が有効だと service account / role / rolebindings が必要 (role は namespace/deployments/pod の create権限....)
             // nodeCount = replicas
 
-            // create namespace
-            _namespaceManifest = KubernetesManifest.GetNamespace(_ns);
-            if (!await _kubeapi.ExistsNamespaceAsync(_ns))
-            {
-                _ = await _kubeapi.CreateNamespaceAsync(_ns, _namespaceManifest, cancellationToken);
-            }
+            //// create namespace
+            //_namespaceManifest = KubernetesManifest.GetNamespace(_ns);
+            //if (!await _kubeapi.ExistsNamespaceAsync(_ns))
+            //{
+            //    _ = await _kubeapi.CreateNamespaceAsync(_ns, _namespaceManifest, cancellationToken);
+            //}
 
             // create deployment
-            _deploymentManifest = KubernetesManifest.GetDeployment(_name, "cysharp/dframe-worker", "0.1.0", options.Host, nodeCount);
+            _deploymentManifest = KubernetesManifest.GetDeployment(_name, "431046970529.dkr.ecr.ap-northeast-1.amazonaws.com/dframe-k8s", "0.1.0", $"{masterSvc}.{masterNamespace}.svc.cluster.local", imagePullPolicy, imagePullSecret, nodeCount);
+            Console.WriteLine(_deploymentManifest);
             _ = await _kubeapi.CreateDeploymentAsync(_ns, _deploymentManifest, cancellationToken);
 
             // wait kubernetes deployments done.
             var deployresult = await _kubeapi.GetDeploymentAsync(_ns, _name);
-            var deploy = _yamlDeserializer.Deserialize<KubernetesDeploymentMetadata>(deployresult);
-            Console.WriteLine($"deployment create. {deploy.metadata.@namespace}/{deploy.metadata.name}");
+            Console.WriteLine(deployresult);
+            Console.WriteLine($"deployment create. {_ns}/{_name}");
         }
 
         public async ValueTask DisposeAsync()
         {
             // delete kubernetes deployments. namespace は master を含むので残す。
-            if (await _kubeapi.ExistsDeploymentAsync(_ns, _name))
+            if (!preserveWorker && await _kubeapi.ExistsDeploymentAsync(_ns, _name))
             {
                 var delete = await _kubeapi.DeleteDeploymentAsync(_ns, _name);
                 Console.WriteLine($"deployment delete. {_ns}/{_name}");
